Move configuration properties out of regular configuration
